fix warning about int vs. float in BlindedState.cpp
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@2083 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ DarkMod/AI/States/BlindedState.cpp
@@ -51,7 +51,7 @@ void BlindedState::Init(idAI* owner)
 		TaskPtr(new SingleBarkTask("snd_blinded"))
 	);
 
-	_endTime = gameLocal.time + 4000 + gameLocal.random.RandomFloat() * 2000;
+	_endTime = gameLocal.time + 4000 + static_cast<int>(gameLocal.random.RandomFloat() * 2000);
 
 	// Set alert level a little bit below combat
 	if(owner->AI_AlertLevel < owner->thresh_5 - 1)
